Update runtime to 5.15-21.08
The 5.15 version of the KDE Runtime is based on the 20.08 version of the Freedesktop Runtime and will stay as-is to keep compatibility.
The 5.15-21.08 version of the KDE Runtime is based on the 21.08 Freedesktop one. The Qt/KDE Libraries are mostly similar between the two runtimes.

This change is mostly maintenance to keep the base runtime updated.