Update to QGIS 3.34.0
While at it, rebase to use the PyQt base app. It's hell to maintain PyQt
and I'm ever grateful for whoever decided to take it over generically.

As a result of this, we drop QtWebKit support, which is dead anyway and
get QtWebEngine instead.

In the process, a number of dependencies get updates.